REFACTOR: 캐시 유틸 함수 공통 모듈로 추출 (코드리뷰 Warning 수정)
formatAmount, getAmountColor, getIconBgColor를 cash-utils.ts로 추출하여
index.tsx, history.tsx 간 중복 제거
