+++ apps/client/src/routes/_auth/my/cash/history.tsx
@@ -5,6 +5,7 @@ import tw from 'tailwind-styled-components';
 
 import SubLayout from '../../../../components/layout/SubLayout';
 import { trpc } from '@/shared';
+import { formatAmount, getAmountColor, getIconBgColor } from './cash-utils';
 
 const LIMIT = 20;
 
@@ -105,26 +106,6 @@ function TypeIcon({ type }: { type: string }) {
   if (type === 'PURCHASE') return <ArrowUpCircle size={size} />;
   if (type === 'REFUND') return <RotateCcw size={size} />;
   return <ArrowDownCircle size={size} />;
-}
-
-function formatAmount(type: string, amount: number): string {
-  const formatted = amount.toLocaleString();
-  if (type === 'CHARGE' || type === 'REFUND') return `+${formatted}원`;
-  return `-${formatted}원`;
-}
-
-function getAmountColor(type: string): string {
-  if (type === 'CHARGE') return 'text-green-600';
-  if (type === 'PURCHASE') return 'text-red-500';
-  if (type === 'REFUND') return 'text-blue-600';
-  return 'text-gray-900';
-}
-
-function getIconBgColor(type: string): string {
-  if (type === 'CHARGE') return 'bg-green-50 text-green-600';
-  if (type === 'PURCHASE') return 'bg-red-50 text-red-500';
-  if (type === 'REFUND') return 'bg-blue-50 text-blue-600';
-  return 'bg-gray-50 text-gray-600';
 }
 
 // Styled Components
